Rename "group name" into "cluster name"
Adjusted groupName and some other configuration parts
to the new changes in the 4.0 API.
