+++ enterprise/pcf-tls-integration/hazelcast.json
@@ -8,7 +8,7 @@
     "updateInterval": 2
   },
   "properties": {},
-  "groupConfig": {
+  "clusterConfig": {
     "name": "dev",
     "password": "dev-pass"
   },
