+++ enterprise/wan-replication/pom.xml
@@ -48,10 +48,5 @@
             <artifactId>commons-lang</artifactId>
             <version>2.6</version>
         </dependency>
-        <dependency>
-            <groupId>com.hazelcast</groupId>
-            <artifactId>hazelcast-solace</artifactId>
-            <version>1.0.0</version>
-        </dependency>
     </dependencies>
 </project>
