+++ enterprise/wan-replication/src/main/java/com/hazelcast/wan/batch/cache/EnterpriseCacheWanReplicationClusterB.java
@@ -102,7 +102,7 @@ public class EnterpriseCacheWanReplicationClusterB {
     private Config getConfigClusterB() {
         Config config = new Config();
         config.setLicenseKey(ENTERPRISE_LICENSE_KEY);
-        config.getGroupConfig().setName("clusterB").setPassword("clusterB-pass");
+        config.setClusterName("clusterB").setClusterPassword("clusterB-pass");
         config.getNetworkConfig().getJoin().getMulticastConfig().setEnabled(false);
         config.getNetworkConfig().getJoin().getTcpIpConfig().setEnabled(true).addMember("127.0.0.1:5702");
         config.setInstanceName("clusterB");
