+++ enterprise/wan-replication/src/main/java/com/hazelcast/wan/batch/map/EnterpriseMapWanReplicationClusterB.java
@@ -3,7 +3,7 @@ package com.hazelcast.wan.batch.map;
 import com.hazelcast.config.Config;
 import com.hazelcast.core.Hazelcast;
 import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.core.IMap;
+import com.hazelcast.map.IMap;
 
 import java.util.Scanner;
 
@@ -61,7 +61,7 @@ public class EnterpriseMapWanReplicationClusterB {
     private static Config getConfigClusterB() {
         Config config = new Config();
         config.setLicenseKey(ENTERPRISE_LICENSE_KEY);
-        config.getGroupConfig().setName("clusterB").setPassword("clusterB-pass");
+        config.setClusterName("clusterB").setClusterPassword("clusterB-pass");
         config.getNetworkConfig().getJoin().getMulticastConfig().setEnabled(false);
         config.getNetworkConfig().getJoin().getTcpIpConfig().setEnabled(true).addMember("127.0.0.1:5702");
         return config;
