+++ hazelcast-integration/pcf-integration/src/main/java/com/hazelcast/pcf/integration/Application.java
@@ -33,12 +33,12 @@ public class Application {
         List hazelcast = (List) json.get("hazelcast");
         Map map = (Map) hazelcast.get(0);
         Map credentials = (Map) map.get("credentials");
-        String groupName = (String) credentials.get("group_name");
-        String groupPassword = (String) credentials.get("group_pass");
+        String clusterName = (String) credentials.get("cluster_name");
+        String clusterPassword = (String) credentials.get("cluster_pass");
         List<String> members = (List<String>) credentials.get("members");
 
         clientConfig = new ClientConfig();
-        clientConfig.setClusterName(groupName).setClusterPassword(groupPassword);
+        clientConfig.setClusterName(clusterName).setClusterPassword(clusterPassword);
         ClientNetworkConfig networkConfig = clientConfig.getNetworkConfig();
         for (String member : members) {
             networkConfig.addAddress(member.replace('"', ' ').trim());
